services/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "AI service for Bomberman Online"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "nats-py>=2.10.0",
    "redis[hiredis]>=6.0.0",
    "redis-dict>=3.2.0",
    "gymnasium[box2d]>=1.1.1",
    "stable-baselines3[extra]>=2.6.0",
    "tensorboard>=2.19.0",
    "python-json-logger",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.9.0",
    "numpy>=2.1.2",
    "py-consul>=1.6.0",
    "aioprometheus[starlette]>=23.12.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.62.0",
    "protobuf>=6.33.4",
]

[dependency-groups]
dev=[
    "ty>=0.0.16",
    "ruff>=0.15.0",
    "pytest>=9.0.2",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/rocm6.3"]
